Move player logic back to clips
